Added explicit compiler to HOOTL.cmake (#53)
The build on the main branch fails to build HOOTLTest on MacOS due to
the default "cc" compiler being used (alias for clang, as far as I can
tell). This PR just sets the C compiler to gcc instead in HOOTL.cmake.

To have this change work, you need to clean rebuild HOOTLTest.

---------

Signed-off-by: Aagrim Hoysal <36976269+ahoysal@users.noreply.github.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Daniel Hansen <105574022+dchansen06@users.noreply.github.com>