G4 Peripheral Testing Project (#103)
# G4 Peripheral Testing Project

## Problem and Scope
We will eventually have to test all of our peripheral abstractions on a
board, yet not all existing projects have a board configured with the
correct peripherals. This could potentially slow down the testing
process for peripheral abstraction, as what may end up happening is that
separate groups spend time on CubeMX simply trying to configure a board
to be able to test their own peripheral.

## Description
G4PERTESTING is a new project for a STM32G4 that has support for SPI,
CAN, TIM, GPIO, DMA (only for ADC), I2C, USART (more like UART
actually), and ADC.

The exact pins should be listed in G4 Peripheral Testing under Misc
Resources in the Gaucho Racing Wiki, or you can just open the .ioc file
that should be included with the project folder.

## Gotchas and Limitations
As hinted above, the peripheral configurations aren't exactly perfect.
USART only has support for asynchronous communication (which is just
UART) and DMA isn't configured for all of the peripherals. In the
future, we may have to boot up CubeMX again and properly configure those
peripherals, but for the most part everything else should have
sufficient support. Shoot me if they don't though.

## Testing

- [X] CMake Builds
- [x] HOOTL testing

### Testing Details
Without the actual peripheral abstractions to work with, there's nothing
really to test other than making sure you can actually get it to debug
properly (which it does).

## Larger Impact
Hopefully this makes peripheral testing a lot easier in the oncoming
weeks when they start being ready to test.

## Additional Context and Ticket

---------

Signed-off-by: Bailey Say <baileyasay@gmail.com>
Co-authored-by: Daniel Hansen <dchansen06@gmail.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Daniel Hansen <105574022+dchansen06@users.noreply.github.com>