Remove Unused STM32L4 CMSIS Header Files and Duplicated Startup Files (#114)
# Removed Unused STM32L4 CMSIS Components

## Problem and Scope
CMSIS was imported for `STM32L476xG` with all of the header files and
startup assembly files in it for the `STM32L4` family, approximately 91%
larger than it needed to be (34 MB vs 3 MB)

## Description
Remove unused header files in the STM32L4 CMSIS library and unused
startup assembly files. Removed duplicate files in CMSIS that were in
`CompileDependencies`.

## Gotchas and Limitations
None

## Testing

- [x] HOOTL testing
- [ ] HITL testing
- [x] Human tested

### Testing Details
L4BLINKY blinks as expected (only @JayWango has one to test)

## Larger Impact
None, the removed header files would never be used considering the
platform in question is explicitly labeled the `STM32L476xG`. The
removed startup assembly files should definitely not be used as they are
for different chipset. Deduplicates the different CMSIS files in
`CompileDependencies`.

## Additional Context and Ticket
Removes ~31 MB.