Correct `CMAKE_BUILD_TYPE` usage (#149)
# Correct CMAKE_BUILD_TYPE usage

## Problem and Scope
CMAKE_BUILD_TYPE has a standard list of appropriate values that does not
include "Test," which the entire monorepo relies on currently.

## Description
Will introduce a new way to check preset state that does not obfuscate
the use of CMAKE_BUILD_TYPE.

## Gotchas and Limitations
All future projects should be mindful of this change, as this changes
the way projects should introduce their tests.

## Testing

- [x] HOOTL testing
- [ ] HITL testing
- [ ] Human tested

### Testing Details
Tested that compilations work for all preset.

## Larger Impact
All future projects should be mindful of this change, as this changes
the way projects should introduce their tests.

## Additional Context and Ticket
https://github.com/Gaucho-Racing/Firmware/issues/85

---------

Co-authored-by: Daniel Hansen <dchansen06@gmail.com>